+++ backend/src/auth/auth.service.ts
@@ -1,6 +1,9 @@
 import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 import { JwtService } from '@nestjs/jwt';
+import { User } from '@prisma/client';
+import { authenticator } from 'otplib';
+import { UsersService } from 'src/users/users.service';
 import { PrismaService } from '../prisma/prisma.service';
 import { SignupUser } from './interfaces/signup-user.interface';
 
@@ -9,7 +12,8 @@ export class AuthService {
   constructor(
     private readonly prisma: PrismaService,
     private readonly config: ConfigService,
-    private readonly jwt: JwtService
+    private readonly jwt: JwtService,
+    private readonly usersService: UsersService
   ) {}
 
   async login(
@@ -36,14 +40,41 @@ export class AuthService {
   async generateJwt(
     id: string,
     name: string,
-    isTwoFactorAuthenticated = false
+    isOtpValid = false
   ): Promise<{ accessToken: string }> {
-    const payload = { id, name, isTwoFactorAuthenticated };
+    const payload = { id, name, isOtpValid };
     const accessToken = await this.jwt.signAsync(payload, {
       expiresIn: '1d',
       secret: this.config.get('JWT_SECRET') as string,
     });
 
     return { accessToken };
+  }
+
+  async generateTwoFactorAuthSecret(user: User): Promise<{
+    otpAuthUrl: string;
+  }> {
+    const secret = authenticator.generateSecret();
+
+    const otpAuthUrl = authenticator.keyuri(
+      user.name,
+      this.config.get<string>('TWO_FACTOR_AUTHENTICATION_APP_NAME') as string,
+      secret
+    );
+
+    await this.usersService.update(user.id, {
+      twoFactorAuthSecret: secret,
+    });
+
+    return {
+      otpAuthUrl,
+    };
+  }
+
+  isTwoFactorAuthCodeValid(twoFactorAuthCode: string, user: User): boolean {
+    return authenticator.verify({
+      token: twoFactorAuthCode,
+      secret: user.twoFactorAuthSecret as string,
+    });
   }
 }
